refactor(plugin): colocate 4 cfg(test) modules with their target submodules (#1092)
Move every section-specific test module into the limited/ or full/
submodule whose code it exercises; only the cross-cutting
gpu_lim_tier1_wire_format_tests harness (~1,380 LoC of null-handle
sweeps across every Limited callback) stays in parent
gpu_context/mod.rs.

* full/mod.rs picks up gpu_full_access_vtable_tests (~570 LoC) — the
  FullAccess static layout-version + drop_handle + create_*
  null-scope-token sweep.
* limited/escalate.rs picks up gpu_lim_escalate_vtable_tests
  (~330 LoC) — the Phase C3 escalate_begin/end round-trip + serial-
  gated NVIDIA dual-VkDevice safety harness.
* limited/texture.rs picks up gpu_lim_texture_native_dma_buf_fd_tests
  (~25 LoC) — the Phase F null-handle -1-sentinel guard.
* limited/video_source_timeline.rs picks up
  gpu_lim_video_source_timeline_semaphore_tests (~70 LoC) — v12 set/
  clear + v14 host_video_source_timeline_arc null-handle guards.

Each colocated test module imports the parent
HOST_GPU_CONTEXT_LIMITED/FULL_ACCESS_VTABLE statics explicitly via
super::super::super::* rather than the old super::* glob — the
relocation loses the parent file's wildcard scope but the explicit
imports are clearer about what the test exercises. The escalate
tests cross both vtable sides and bring in both statics.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored on a clean build (cargo's
incremental cache occasionally trips a SIGSEGV across module
restructures; a cargo clean -p streamlib-engine before rerunning
resolves it deterministically).
