+++ libs/streamlib-engine/src/core/plugin/host_services/gpu_context/limited/escalate.rs
@@ -113,3 +113,337 @@ pub(in crate::core::plugin::host_services) unsafe extern "C" fn host_gpu_lim_esc
         1,
     )
 }
+#[cfg(test)]
+mod gpu_lim_escalate_vtable_tests {
+    //! Tier-1 wire-format + round-trip tests for C3's escalate_begin
+    //! and escalate_end vtable entries.
+    //!
+    //! Tests that construct a real `GpuContext` carry `#[serial]` to
+    //! prevent the NVIDIA Linux dual-`VkDevice` SIGSEGV
+    //! (`docs/learnings/nvidia-dual-vulkan-device-crash.md`) when run
+    //! against other VkDevice-creating tests in the workspace lib
+    //! suite.
+
+    use std::ffi::c_void;
+    use std::sync::Arc;
+
+    use super::super::super::{
+        HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE, HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE,
+    };
+    use serial_test::serial;
+
+    fn make_err_buf() -> ([u8; 256], usize) {
+        ([0u8; 256], 0usize)
+    }
+
+    fn err_buf_as_str(buf: &[u8], len: usize) -> &str {
+        std::str::from_utf8(&buf[..len]).expect("UTF-8")
+    }
+
+    /// Build a host-mode gpu_handle (the `Box<Arc<GpuContext>>`-shaped
+    /// pointer that `GpuContextLimitedAccess::new` produces) so the
+    /// `escalate_begin` callback can run end-to-end against a real
+    /// `Arc<GpuContext>`. Skips when no GPU device is available.
+    fn make_host_handle() -> Option<(*const c_void, Arc<crate::core::context::GpuContext>)> {
+        let gpu = crate::core::context::GpuContext::init_for_platform().ok()?;
+        let arc = Arc::new(gpu);
+        let boxed: Box<Arc<crate::core::context::GpuContext>> = Box::new(Arc::clone(&arc));
+        let handle = Box::into_raw(boxed) as *const c_void;
+        Some((handle, arc))
+    }
+
+    /// Free a host_handle minted by `make_host_handle` — pairs with
+    /// the `Box::into_raw`.
+    unsafe fn free_host_handle(handle: *const c_void) {
+        let _ = unsafe {
+            Box::from_raw(handle as *mut Arc<crate::core::context::GpuContext>)
+        };
+    }
+
+    #[test]
+    fn escalate_begin_returns_error_on_null_gpu_handle() {
+        let (mut buf, mut len) = make_err_buf();
+        let mut token: *const c_void = std::ptr::null();
+        let rc = unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
+                std::ptr::null(),
+                &mut token,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            )
+        };
+        assert_eq!(rc, 1);
+        let msg = err_buf_as_str(&buf, len);
+        assert!(msg.contains("escalate_begin: null gpu handle"), "got: {msg}");
+        assert!(token.is_null(), "scope token must not be written on error");
+    }
+
+    #[test]
+    #[serial]
+    fn escalate_begin_returns_error_on_null_out_param() {
+        let Some((handle, _arc)) = make_host_handle() else {
+            tracing::warn!(
+                target: "streamlib::tests::escalate_vtable",
+                "skipping escalate_begin null-out test: no GPU device"
+            );
+            return;
+        };
+        let (mut buf, mut len) = make_err_buf();
+        let rc = unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
+                handle,
+                std::ptr::null_mut(),
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            )
+        };
+        assert_eq!(rc, 1);
+        let msg = err_buf_as_str(&buf, len);
+        assert!(
+            msg.contains("escalate_begin: null out_scope_token"),
+            "got: {msg}"
+        );
+        unsafe { free_host_handle(handle) };
+    }
+
+    #[test]
+    fn escalate_end_is_idempotent_for_stale_token() {
+        // escalate_end with a never-issued token is a clean no-op
+        // (returns 0; doesn't release any gate). Documented as
+        // idempotent in the registry.
+        let (mut buf, mut len) = make_err_buf();
+        let rc = unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
+                std::ptr::null(),
+                u64::MAX as *const c_void, // never-issued token
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            )
+        };
+        assert_eq!(rc, 0);
+        assert_eq!(len, 0, "no error message expected for stale token");
+    }
+
+    #[test]
+    #[serial]
+    fn round_trip_begin_then_end_releases_gate() {
+        let Some((handle, _arc)) = make_host_handle() else {
+            tracing::warn!(
+                target: "streamlib::tests::escalate_vtable",
+                "skipping round-trip test: no GPU device"
+            );
+            return;
+        };
+
+        let (mut buf, mut len) = make_err_buf();
+        let mut token: *const c_void = std::ptr::null();
+        let begin_rc = unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
+                handle,
+                &mut token,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            )
+        };
+        assert_eq!(begin_rc, 0);
+        assert!(!token.is_null(), "scope token must be written on success");
+
+        let end_rc = unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
+                handle,
+                token,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            )
+        };
+        assert_eq!(end_rc, 0);
+
+        // Begin again on the same handle — gate must have been
+        // released, so this succeeds without blocking. (If the gate
+        // hadn't released, this would deadlock.)
+        let mut token2: *const c_void = std::ptr::null();
+        let begin2_rc = unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
+                handle,
+                &mut token2,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            )
+        };
+        assert_eq!(begin2_rc, 0);
+        assert!(!token2.is_null());
+        assert_ne!(token, token2, "tokens must be unique per begin call");
+
+        let _ = unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
+                handle,
+                token2,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            )
+        };
+        unsafe { free_host_handle(handle) };
+    }
+
+    #[test]
+    #[serial]
+    fn full_access_callback_with_valid_token_resolves_scope() {
+        // End-to-end: begin a scope, get a valid token, invoke a
+        // FullAccess vtable callback with the token + a valid
+        // descriptor. The callback's scope-token lookup must succeed
+        // (no "invalid escalate scope" error). The actual allocation
+        // may succeed or fail depending on the Vulkan environment
+        // (render-target DMA-BUF availability, EGL DRM modifier
+        // probe), but EITHER outcome proves the scope lookup passed:
+        // a success returns rc=0 with `out_texture` populated; a
+        // failure returns rc=1 with an error message that does NOT
+        // contain "invalid escalate scope".
+        //
+        // (Mentally revert `with_full_scope_or_err` to always return
+        // None — this test fails because the error message would
+        // then contain "invalid escalate scope".)
+        let Some((handle, _arc)) = make_host_handle() else {
+            tracing::warn!(
+                target: "streamlib::tests::escalate_vtable",
+                "skipping valid-token test: no GPU device"
+            );
+            return;
+        };
+
+        let (mut buf, mut len) = make_err_buf();
+        let mut token: *const c_void = std::ptr::null();
+        unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
+                handle,
+                &mut token,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            );
+        }
+        assert!(!token.is_null());
+
+        let mut out: crate::core::rhi::texture::Texture =
+            unsafe { std::mem::zeroed() };
+        let mut buf2 = [0u8; 256];
+        let mut len2 = 0usize;
+        let rc = unsafe {
+            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE
+                .acquire_render_target_dma_buf_image)(
+                token,
+                64,
+                64,
+                0, // Rgba8Unorm — valid format; forces scope lookup to run
+                &mut out as *mut _ as *mut c_void,
+                buf2.as_mut_ptr(),
+                buf2.len(),
+                &mut len2,
+            )
+        };
+
+        if rc != 0 {
+            // Allocation failed for an environment reason; assert the
+            // failure was NOT a scope-lookup miss.
+            let msg = err_buf_as_str(&buf2, len2);
+            assert!(
+                !msg.contains("invalid escalate scope"),
+                "scope-token lookup must succeed inside an active \
+                 scope; got: {msg}"
+            );
+        } else {
+            // Allocation succeeded — definitively proves scope lookup
+            // worked. The Texture in `out` owns a live handle; its
+            // Drop will fire the vtable's drop_texture as the test
+            // returns.
+            assert!(!out.handle.is_null(), "out_texture handle populated");
+            // SAFETY: `out` was overwritten by `ptr::write` from the
+            // callback with a valid Texture; let its normal Drop run
+            // to release the underlying handle via the vtable.
+        }
+
+        // Clean up the scope.
+        unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
+                handle,
+                token,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            );
+        }
+        unsafe { free_host_handle(handle) };
+    }
+
+    #[test]
+    #[serial]
+    fn full_access_callback_fails_after_escalate_end() {
+        // Closes the scope-token validation loop: a token used after
+        // escalate_end fires returns the InvalidEscalateScope error
+        // (matches the "calls after escalate_end return
+        // InvalidEscalateScope" exit criterion).
+        let Some((handle, _arc)) = make_host_handle() else {
+            tracing::warn!(
+                target: "streamlib::tests::escalate_vtable",
+                "skipping post-end test: no GPU device"
+            );
+            return;
+        };
+
+        let (mut buf, mut len) = make_err_buf();
+        let mut token: *const c_void = std::ptr::null();
+        unsafe {
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
+                handle,
+                &mut token,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            );
+            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
+                handle,
+                token,
+                buf.as_mut_ptr(),
+                buf.len(),
+                &mut len,
+            );
+        }
+
+        // Token is now stale — using it on any FullAccess callback
+        // returns "invalid escalate scope".
+        let mut out: crate::core::rhi::texture::Texture =
+            unsafe { std::mem::zeroed() };
+        let mut buf2 = [0u8; 256];
+        let mut len2 = 0usize;
+        let rc = unsafe {
+            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE
+                .acquire_render_target_dma_buf_image)(
+                token,
+                64,
+                64,
+                0, // valid format
+                &mut out as *mut _ as *mut c_void,
+                buf2.as_mut_ptr(),
+                buf2.len(),
+                &mut len2,
+            )
+        };
+        assert_eq!(rc, 1);
+        let msg = err_buf_as_str(&buf2, len2);
+        assert!(
+            msg.contains(
+                "acquire_render_target_dma_buf_image: invalid escalate scope"
+            ),
+            "got: {msg}"
+        );
+
+        unsafe { free_host_handle(handle) };
+    }
+}
+
