+++ libs/streamlib-engine/src/core/plugin/host_services/gpu_context/mod.rs
@@ -299,989 +299,6 @@ pub fn host_gpu_context_limited_access_vtable() -> *const GpuContextLimitedAcces
     }
 }
 #[cfg(test)]
-mod gpu_full_access_vtable_tests {
-    use super::*;
-    use streamlib_plugin_abi::{
-        ComputeKernelDescriptorRepr, GraphicsKernelDescriptorRepr,
-        RayTracingKernelDescriptorRepr,
-    };
-
-    fn make_err_buf() -> ([u8; 256], usize) {
-        ([0u8; 256], 0usize)
-    }
-
-    fn err_buf_as_str(buf: &[u8], len: usize) -> &str {
-        std::str::from_utf8(&buf[..len]).expect("UTF-8")
-    }
-
-    #[test]
-    fn drop_handle_handles_null_no_crash() {
-        // Null handle is documented as a no-op; this just exercises
-        // the early-return guard.
-        unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.drop_handle)(std::ptr::null());
-        }
-    }
-
-    #[test]
-    fn create_compute_kernel_returns_error_on_null_scope_token() {
-        // Post-C3: gpu_handle is interpreted as a scope_token; a null
-        // pointer corresponds to scope_token = 0, which is reserved as
-        // "never issued" — `with_scope` returns None and the callback
-        // returns an "invalid escalate scope" error.
-        let (mut buf, mut len) = make_err_buf();
-        let mut out: *const c_void = std::ptr::null();
-        let bindings_buf: [streamlib_plugin_abi::ComputeBindingSpecRepr; 0] = [];
-        let repr = ComputeKernelDescriptorRepr {
-            label_ptr: "test".as_ptr(),
-            label_len: 4,
-            spv_ptr: std::ptr::null(),
-            spv_len: 0,
-            bindings_ptr: bindings_buf.as_ptr(),
-            bindings_len: 0,
-            push_constant_size: 0,
-            _reserved_padding: 0,
-        };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.create_compute_kernel)(
-                std::ptr::null(),
-                &repr,
-                &mut out,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("create_compute_kernel: invalid escalate scope"),
-            "got: {msg}"
-        );
-        assert!(out.is_null(), "out_kernel must not be written on error");
-    }
-
-    #[test]
-    fn create_graphics_kernel_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut out: *const c_void = std::ptr::null();
-        let repr: GraphicsKernelDescriptorRepr = unsafe { std::mem::zeroed() };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.create_graphics_kernel)(
-                std::ptr::null(),
-                &repr,
-                &mut out,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("create_graphics_kernel: invalid escalate scope"),
-            "got: {msg}"
-        );
-        assert!(out.is_null());
-    }
-
-    #[test]
-    fn create_ray_tracing_kernel_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut out: *const c_void = std::ptr::null();
-        let repr: RayTracingKernelDescriptorRepr = unsafe { std::mem::zeroed() };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.create_ray_tracing_kernel)(
-                std::ptr::null(),
-                &repr,
-                &mut out,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("create_ray_tracing_kernel: invalid escalate scope"),
-            "got: {msg}"
-        );
-        assert!(out.is_null());
-    }
-
-    #[test]
-    fn create_texture_ring_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut out: *const c_void = std::ptr::null();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.create_texture_ring)(
-                std::ptr::null(),
-                64,
-                64,
-                0, // Rgba8Unorm
-                0, // no usage bits
-                2,
-                &mut out,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("create_texture_ring: invalid escalate scope"),
-            "got: {msg}"
-        );
-        assert!(out.is_null());
-    }
-
-    #[test]
-    fn acquire_render_target_dma_buf_image_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut out: crate::core::rhi::texture::Texture =
-            unsafe { std::mem::zeroed() };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE
-                .acquire_render_target_dma_buf_image)(
-                std::ptr::null(),
-                64,
-                64,
-                0, // Rgba8Unorm
-                &mut out as *mut _ as *mut c_void,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains(
-                "acquire_render_target_dma_buf_image: invalid escalate scope"
-            ),
-            "got: {msg}"
-        );
-    }
-
-    #[test]
-    fn acquire_render_target_dma_buf_image_returns_error_on_invalid_format() {
-        // Even with an invalid format, the null scope-token check would
-        // run after the format decode — so feeding a token of 0 (which
-        // would later fail scope lookup) but an invalid format ensures
-        // the format-validation path fires.
-        let (mut buf, mut len) = make_err_buf();
-        let mut out: crate::core::rhi::texture::Texture =
-            unsafe { std::mem::zeroed() };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE
-                .acquire_render_target_dma_buf_image)(
-                std::ptr::null(),
-                64,
-                64,
-                99, // invalid format_raw
-                &mut out as *mut _ as *mut c_void,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains(
-                "acquire_render_target_dma_buf_image: invalid format_raw"
-            ),
-            "got: {msg}"
-        );
-    }
-
-    // ============================================================================
-    // Phase D (#906) — tier-1 wire-format tests for the 9 new FullAccess slots
-    // ============================================================================
-
-    #[test]
-    fn wait_device_idle_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.wait_device_idle)(
-                std::ptr::null(),
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("wait_device_idle: invalid escalate scope"),
-            "got: {msg}"
-        );
-    }
-
-    #[test]
-    fn acquire_output_texture_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut id_buf = [0u8; 256];
-        let mut id_len: usize = 0;
-        let mut out: crate::core::rhi::texture::Texture =
-            unsafe { std::mem::zeroed() };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.acquire_output_texture)(
-                std::ptr::null(),
-                64,
-                64,
-                0,
-                id_buf.as_mut_ptr(),
-                id_buf.len(),
-                &mut id_len,
-                &mut out as *mut _ as *mut c_void,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("acquire_output_texture: invalid escalate scope"),
-            "got: {msg}"
-        );
-    }
-
-    #[test]
-    fn acquire_output_texture_returns_error_on_invalid_format() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut id_buf = [0u8; 256];
-        let mut id_len: usize = 0;
-        let mut out: crate::core::rhi::texture::Texture =
-            unsafe { std::mem::zeroed() };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.acquire_output_texture)(
-                std::ptr::null(),
-                64,
-                64,
-                99,
-                id_buf.as_mut_ptr(),
-                id_buf.len(),
-                &mut id_len,
-                &mut out as *mut _ as *mut c_void,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("acquire_output_texture: invalid format_raw"),
-            "got: {msg}"
-        );
-    }
-
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn upload_pixel_buffer_as_texture_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        // We pass non-null surface_id + a "borrowed" PixelBuffer placeholder
-        // through the null-pointer guard; the scope-token check then fires
-        // because the token is null/zero.
-        let sid = b"abc";
-        let pb: crate::core::rhi::PixelBuffer = unsafe { std::mem::zeroed() };
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.upload_pixel_buffer_as_texture)(
-                std::ptr::null(),
-                sid.as_ptr(),
-                sid.len(),
-                &pb as *const _ as *const c_void,
-                64,
-                64,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        // Leak the zeroed PixelBuffer to avoid running its (cdylib-mode)
-        // Drop on a null handle — that would dispatch through a null
-        // vtable. The null-handle Drop guard short-circuits, but
-        // mem::forget makes the intent explicit.
-        std::mem::forget(pb);
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("upload_pixel_buffer_as_texture: invalid escalate scope"),
-            "got: {msg}"
-        );
-    }
-
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn color_converter_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut out: *const c_void = std::ptr::null();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.color_converter)(
-                std::ptr::null(),
-                0, // src
-                0, // dst
-                &mut out,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("color_converter: invalid escalate scope"),
-            "got: {msg}"
-        );
-        assert!(out.is_null());
-    }
-
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn create_command_recorder_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let label = b"test_recorder";
-        let mut out: std::mem::MaybeUninit<crate::vulkan::rhi::RhiCommandRecorder> =
-            std::mem::MaybeUninit::uninit();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.create_command_recorder)(
-                std::ptr::null(),
-                label.as_ptr(),
-                label.len(),
-                out.as_mut_ptr() as *mut c_void,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("create_command_recorder: invalid escalate scope"),
-            "got: {msg}"
-        );
-    }
-
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn build_triangles_blas_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let label = b"test_blas";
-        let vertices = [0.0f32, 0.0, 0.0];
-        let indices = [0u32, 1, 2];
-        let mut out: *const c_void = std::ptr::null();
-        let mut out_device_address: u64 = 0;
-        let mut out_storage_size: u64 = 0;
-        let mut out_kind: u32 = 0;
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.build_triangles_blas)(
-                std::ptr::null(),
-                label.as_ptr(),
-                label.len(),
-                vertices.as_ptr(),
-                vertices.len(),
-                indices.as_ptr(),
-                indices.len(),
-                &mut out,
-                &mut out_device_address as *mut u64,
-                &mut out_storage_size as *mut u64,
-                &mut out_kind as *mut u32,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("build_triangles_blas: invalid escalate scope"),
-            "got: {msg}"
-        );
-        assert!(out.is_null());
-        // Out-params untouched on failure.
-        assert_eq!(out_device_address, 0);
-        assert_eq!(out_storage_size, 0);
-        assert_eq!(out_kind, 0);
-    }
-
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn build_tlas_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let label = b"test_tlas";
-        let mut out: *const c_void = std::ptr::null();
-        let mut out_device_address: u64 = 0;
-        let mut out_storage_size: u64 = 0;
-        let mut out_kind: u32 = 0;
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.build_tlas)(
-                std::ptr::null(),
-                label.as_ptr(),
-                label.len(),
-                std::ptr::null(),
-                0,
-                &mut out,
-                &mut out_device_address as *mut u64,
-                &mut out_storage_size as *mut u64,
-                &mut out_kind as *mut u32,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("build_tlas: invalid escalate scope"),
-            "got: {msg}"
-        );
-        assert!(out.is_null());
-        // Out-params untouched on failure.
-        assert_eq!(out_device_address, 0);
-        assert_eq!(out_storage_size, 0);
-        assert_eq!(out_kind, 0);
-    }
-
-    #[test]
-    #[cfg(target_os = "linux")]
-    fn supports_ray_tracing_pipeline_returns_negative_one_on_null_scope_token() {
-        // Returns -1 for "invalid scope token" (since 1/0 are valid yes/no
-        // bool returns). The error message goes to err_buf.
-        let (mut buf, mut len) = make_err_buf();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.supports_ray_tracing_pipeline)(
-                std::ptr::null(),
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, -1, "null scope token must return -1, got {rc}");
-    }
-
-    #[test]
-    fn check_in_surface_returns_error_on_null_scope_token() {
-        let (mut buf, mut len) = make_err_buf();
-        let pb: crate::core::rhi::PixelBuffer = unsafe { std::mem::zeroed() };
-        let mut id_buf = [0u8; 256];
-        let mut id_len: usize = 0;
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.check_in_surface)(
-                std::ptr::null(),
-                &pb as *const _ as *const c_void,
-                id_buf.as_mut_ptr(),
-                id_buf.len(),
-                &mut id_len,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        std::mem::forget(pb);
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("check_in_surface: invalid escalate scope"),
-            "got: {msg}"
-        );
-    }
-
-    #[test]
-    fn vtable_layout_version_matches_constant() {
-        assert_eq!(
-            HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.layout_version,
-            streamlib_plugin_abi::GPU_CONTEXT_FULL_ACCESS_VTABLE_LAYOUT_VERSION
-        );
-    }
-
-    // v9 slot: host_vulkan_device_arc takes a scope token (not an
-    // err-buf). The null-token case bottoms out in
-    // `with_scope(0, ...) → None`, so the callback returns a null
-    // pointer. Mental-revert: stub the callback body to call
-    // `Arc::into_raw(...)` directly on a freshly-cloned Arc without
-    // checking the token; this test trips on the resulting non-null
-    // return and the unmatched `from_raw` would Drop the Arc, lowering
-    // the refcount on the host's actual `Arc<HostVulkanDevice>`.
-    #[test]
-    fn host_vulkan_device_arc_returns_null_on_null_token() {
-        let raw = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.host_vulkan_device_arc)(
-                std::ptr::null(),
-            )
-        };
-        assert!(raw.is_null(), "null scope token must yield null pointer");
-    }
-
-    // v10 slot: host_vulkan_texture_arc takes a raw texture handle
-    // (not an err-buf). The null-handle case short-circuits in the
-    // wrapper before any deref, returning a null pointer. Mental-
-    // revert: remove the `if texture_handle.is_null()` guard inside
-    // `host_gpu_full_host_vulkan_texture_arc`; the wrapper would then
-    // UB-deref the null pointer as `*const TextureInner` and the test
-    // runner would SIGSEGV.
-    #[test]
-    fn host_vulkan_texture_arc_returns_null_on_null_handle() {
-        let raw = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE.host_vulkan_texture_arc)(
-                std::ptr::null(),
-            )
-        };
-        assert!(raw.is_null(), "null texture handle must yield null pointer");
-    }
-
-    #[test]
-    fn host_services_for_self_wires_full_access_vtable() {
-        let node = match crate::iceoryx2::Iceoryx2Node::new() {
-            Ok(n) => n,
-            Err(e) => {
-                tracing::warn!(
-                    target: "streamlib::tests::gpu_full_access_vtable",
-                    error = %e,
-                    "skipping host_services_for_self wiring assertion: iceoryx2 init unavailable in this env"
-                );
-                return;
-            }
-        };
-        let services = super::super::runtime_facing::host_services_for_self(&node);
-        assert!(
-            !services.gpu_context_full_access_vtable.is_null(),
-            "host should wire the FullAccess vtable pointer"
-        );
-        let installed_version =
-            unsafe { (*services.gpu_context_full_access_vtable).layout_version };
-        assert_eq!(
-            installed_version,
-            streamlib_plugin_abi::GPU_CONTEXT_FULL_ACCESS_VTABLE_LAYOUT_VERSION
-        );
-    }
-}
-#[cfg(test)]
-mod gpu_lim_escalate_vtable_tests {
-    //! Tier-1 wire-format + round-trip tests for C3's escalate_begin
-    //! and escalate_end vtable entries.
-    //!
-    //! Tests that construct a real `GpuContext` carry `#[serial]` to
-    //! prevent the NVIDIA Linux dual-`VkDevice` SIGSEGV
-    //! (`docs/learnings/nvidia-dual-vulkan-device-crash.md`) when run
-    //! against other VkDevice-creating tests in the workspace lib
-    //! suite.
-
-    use super::*;
-    use serial_test::serial;
-
-    fn make_err_buf() -> ([u8; 256], usize) {
-        ([0u8; 256], 0usize)
-    }
-
-    fn err_buf_as_str(buf: &[u8], len: usize) -> &str {
-        std::str::from_utf8(&buf[..len]).expect("UTF-8")
-    }
-
-    /// Build a host-mode gpu_handle (the `Box<Arc<GpuContext>>`-shaped
-    /// pointer that `GpuContextLimitedAccess::new` produces) so the
-    /// `escalate_begin` callback can run end-to-end against a real
-    /// `Arc<GpuContext>`. Skips when no GPU device is available.
-    fn make_host_handle() -> Option<(*const c_void, Arc<crate::core::context::GpuContext>)> {
-        let gpu = crate::core::context::GpuContext::init_for_platform().ok()?;
-        let arc = Arc::new(gpu);
-        let boxed: Box<Arc<crate::core::context::GpuContext>> = Box::new(Arc::clone(&arc));
-        let handle = Box::into_raw(boxed) as *const c_void;
-        Some((handle, arc))
-    }
-
-    /// Free a host_handle minted by `make_host_handle` — pairs with
-    /// the `Box::into_raw`.
-    unsafe fn free_host_handle(handle: *const c_void) {
-        let _ = unsafe {
-            Box::from_raw(handle as *mut Arc<crate::core::context::GpuContext>)
-        };
-    }
-
-    #[test]
-    fn escalate_begin_returns_error_on_null_gpu_handle() {
-        let (mut buf, mut len) = make_err_buf();
-        let mut token: *const c_void = std::ptr::null();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
-                std::ptr::null(),
-                &mut token,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(msg.contains("escalate_begin: null gpu handle"), "got: {msg}");
-        assert!(token.is_null(), "scope token must not be written on error");
-    }
-
-    #[test]
-    #[serial]
-    fn escalate_begin_returns_error_on_null_out_param() {
-        let Some((handle, _arc)) = make_host_handle() else {
-            tracing::warn!(
-                target: "streamlib::tests::escalate_vtable",
-                "skipping escalate_begin null-out test: no GPU device"
-            );
-            return;
-        };
-        let (mut buf, mut len) = make_err_buf();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
-                handle,
-                std::ptr::null_mut(),
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf, len);
-        assert!(
-            msg.contains("escalate_begin: null out_scope_token"),
-            "got: {msg}"
-        );
-        unsafe { free_host_handle(handle) };
-    }
-
-    #[test]
-    fn escalate_end_is_idempotent_for_stale_token() {
-        // escalate_end with a never-issued token is a clean no-op
-        // (returns 0; doesn't release any gate). Documented as
-        // idempotent in the registry.
-        let (mut buf, mut len) = make_err_buf();
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
-                std::ptr::null(),
-                u64::MAX as *const c_void, // never-issued token
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(rc, 0);
-        assert_eq!(len, 0, "no error message expected for stale token");
-    }
-
-    #[test]
-    #[serial]
-    fn round_trip_begin_then_end_releases_gate() {
-        let Some((handle, _arc)) = make_host_handle() else {
-            tracing::warn!(
-                target: "streamlib::tests::escalate_vtable",
-                "skipping round-trip test: no GPU device"
-            );
-            return;
-        };
-
-        let (mut buf, mut len) = make_err_buf();
-        let mut token: *const c_void = std::ptr::null();
-        let begin_rc = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
-                handle,
-                &mut token,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(begin_rc, 0);
-        assert!(!token.is_null(), "scope token must be written on success");
-
-        let end_rc = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
-                handle,
-                token,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(end_rc, 0);
-
-        // Begin again on the same handle — gate must have been
-        // released, so this succeeds without blocking. (If the gate
-        // hadn't released, this would deadlock.)
-        let mut token2: *const c_void = std::ptr::null();
-        let begin2_rc = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
-                handle,
-                &mut token2,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        assert_eq!(begin2_rc, 0);
-        assert!(!token2.is_null());
-        assert_ne!(token, token2, "tokens must be unique per begin call");
-
-        let _ = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
-                handle,
-                token2,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            )
-        };
-        unsafe { free_host_handle(handle) };
-    }
-
-    #[test]
-    #[serial]
-    fn full_access_callback_with_valid_token_resolves_scope() {
-        // End-to-end: begin a scope, get a valid token, invoke a
-        // FullAccess vtable callback with the token + a valid
-        // descriptor. The callback's scope-token lookup must succeed
-        // (no "invalid escalate scope" error). The actual allocation
-        // may succeed or fail depending on the Vulkan environment
-        // (render-target DMA-BUF availability, EGL DRM modifier
-        // probe), but EITHER outcome proves the scope lookup passed:
-        // a success returns rc=0 with `out_texture` populated; a
-        // failure returns rc=1 with an error message that does NOT
-        // contain "invalid escalate scope".
-        //
-        // (Mentally revert `with_full_scope_or_err` to always return
-        // None — this test fails because the error message would
-        // then contain "invalid escalate scope".)
-        let Some((handle, _arc)) = make_host_handle() else {
-            tracing::warn!(
-                target: "streamlib::tests::escalate_vtable",
-                "skipping valid-token test: no GPU device"
-            );
-            return;
-        };
-
-        let (mut buf, mut len) = make_err_buf();
-        let mut token: *const c_void = std::ptr::null();
-        unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
-                handle,
-                &mut token,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            );
-        }
-        assert!(!token.is_null());
-
-        let mut out: crate::core::rhi::texture::Texture =
-            unsafe { std::mem::zeroed() };
-        let mut buf2 = [0u8; 256];
-        let mut len2 = 0usize;
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE
-                .acquire_render_target_dma_buf_image)(
-                token,
-                64,
-                64,
-                0, // Rgba8Unorm — valid format; forces scope lookup to run
-                &mut out as *mut _ as *mut c_void,
-                buf2.as_mut_ptr(),
-                buf2.len(),
-                &mut len2,
-            )
-        };
-
-        if rc != 0 {
-            // Allocation failed for an environment reason; assert the
-            // failure was NOT a scope-lookup miss.
-            let msg = err_buf_as_str(&buf2, len2);
-            assert!(
-                !msg.contains("invalid escalate scope"),
-                "scope-token lookup must succeed inside an active \
-                 scope; got: {msg}"
-            );
-        } else {
-            // Allocation succeeded — definitively proves scope lookup
-            // worked. The Texture in `out` owns a live handle; its
-            // Drop will fire the vtable's drop_texture as the test
-            // returns.
-            assert!(!out.handle.is_null(), "out_texture handle populated");
-            // SAFETY: `out` was overwritten by `ptr::write` from the
-            // callback with a valid Texture; let its normal Drop run
-            // to release the underlying handle via the vtable.
-        }
-
-        // Clean up the scope.
-        unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
-                handle,
-                token,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            );
-        }
-        unsafe { free_host_handle(handle) };
-    }
-
-    #[test]
-    #[serial]
-    fn full_access_callback_fails_after_escalate_end() {
-        // Closes the scope-token validation loop: a token used after
-        // escalate_end fires returns the InvalidEscalateScope error
-        // (matches the "calls after escalate_end return
-        // InvalidEscalateScope" exit criterion).
-        let Some((handle, _arc)) = make_host_handle() else {
-            tracing::warn!(
-                target: "streamlib::tests::escalate_vtable",
-                "skipping post-end test: no GPU device"
-            );
-            return;
-        };
-
-        let (mut buf, mut len) = make_err_buf();
-        let mut token: *const c_void = std::ptr::null();
-        unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_begin)(
-                handle,
-                &mut token,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            );
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE.escalate_end)(
-                handle,
-                token,
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut len,
-            );
-        }
-
-        // Token is now stale — using it on any FullAccess callback
-        // returns "invalid escalate scope".
-        let mut out: crate::core::rhi::texture::Texture =
-            unsafe { std::mem::zeroed() };
-        let mut buf2 = [0u8; 256];
-        let mut len2 = 0usize;
-        let rc = unsafe {
-            (HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE
-                .acquire_render_target_dma_buf_image)(
-                token,
-                64,
-                64,
-                0, // valid format
-                &mut out as *mut _ as *mut c_void,
-                buf2.as_mut_ptr(),
-                buf2.len(),
-                &mut len2,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = err_buf_as_str(&buf2, len2);
-        assert!(
-            msg.contains(
-                "acquire_render_target_dma_buf_image: invalid escalate scope"
-            ),
-            "got: {msg}"
-        );
-
-        unsafe { free_host_handle(handle) };
-    }
-}
-#[cfg(test)]
-mod gpu_lim_texture_native_dma_buf_fd_tests {
-    //! Tier-1 wire-format test for the Phase F
-    //! `texture_native_dma_buf_fd` slot (#908 / #957). The slot is the
-    //! cross-DSO landing for `Texture::native_handle` on Linux and
-    //! returns the DMA-BUF FD widened to `i64`; sentinel `-1` encodes
-    //! the `Option::None` case. A null texture handle must be a clean
-    //! `-1` (no panic, no UB) — the wrapper short-circuits before any
-    //! cast through `*const TextureInner`.
-
-    use super::*;
-
-    #[test]
-    fn texture_native_dma_buf_fd_returns_minus_one_on_null_handle() {
-        // Null texture_handle is the cdylib-shaped "Texture wasn't
-        // minted yet / was already dropped" case. The slot returns
-        // `-1` (= `Option::None` in the Rust-side wrapper) without
-        // panicking and without touching the null pointer.
-        let fd = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE
-                .texture_native_dma_buf_fd)(std::ptr::null())
-        };
-        assert_eq!(
-            fd, -1,
-            "null texture_handle must produce -1 sentinel (None)"
-        );
-    }
-}
-#[cfg(test)]
-mod gpu_lim_video_source_timeline_semaphore_tests {
-    //! Tier-1 wire-format tests for the v12 (#958)
-    //! `set_video_source_timeline_semaphore` /
-    //! `clear_video_source_timeline_semaphore` slots. Each wrapper
-    //! must short-circuit on null gpu_handle (and `set` on null
-    //! timeline_handle) without panicking and without dereferencing
-    //! the null pointers.
-    //!
-    //! The non-null-handle path is exercised end-to-end by the
-    //! `load_project_dylib_camera_smoke` integration test (which
-    //! holds a real `Arc<HostVulkanTimelineSemaphore>` and is the
-    //! only place a Tier-1 with-handle test could reach without
-    //! constructing a real `GpuContext` here).
-    //!
-    //! Mental-revert: stub the wrapper bodies to
-    //! `unimplemented!()` and these tests trip the underlying
-    //! deref / panic — the wire-format claim regresses.
-    use super::*;
-
-    #[test]
-    fn set_video_source_timeline_is_noop_on_null_gpu_handle() {
-        unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE
-                .set_video_source_timeline_semaphore)(
-                std::ptr::null(),
-                std::ptr::null(),
-            );
-        }
-    }
-
-    // Note: the timeline_handle null guard at host_gpu_lim_set_video_source_timeline_semaphore
-    // line 2078 isn't reachable at tier-1: the first guard
-    // (handle_as_gpu_context) short-circuits on null gpu_handle, and
-    // a non-null garbage gpu_handle would UB-deref before reaching
-    // the timeline check. The guard is exercised end-to-end by
-    // load_project_dylib_camera_smoke (the cdylib camera passes a
-    // valid gpu_handle and a real Arc-borrow timeline_handle).
-
-    #[test]
-    fn clear_video_source_timeline_is_noop_on_null_gpu_handle() {
-        unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE
-                .clear_video_source_timeline_semaphore)(std::ptr::null());
-        }
-    }
-
-    /// v14 slot (#1066): tier-1 wire-format guard. Null `gpu_handle`
-    /// must return null rather than dereferencing the pointer. The
-    /// non-null-handle "slot empty" → null and "slot populated" →
-    /// non-null Arc pointer paths are exercised end-to-end by the
-    /// camera-display cdylib reproducer; a tier-1 unit test for them
-    /// would need a real `GpuContext` instance, which this module
-    /// deliberately avoids constructing.
-    ///
-    /// Mental-revert: stub the wrapper to `unimplemented!()` and
-    /// this test trips the underlying panic — the null-guard
-    /// contract regresses.
-    #[test]
-    fn host_video_source_timeline_arc_returns_null_on_null_gpu_handle() {
-        let raw = unsafe {
-            (HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE
-                .host_video_source_timeline_arc)(std::ptr::null())
-        };
-        assert!(raw.is_null(), "expected null on null gpu_handle");
-    }
-}
-#[cfg(test)]
 mod gpu_lim_tier1_wire_format_tests {
     //! Tier-1 wire-format tests for every callback on
     //! [`GpuContextLimitedAccessVTable`].
